requirements.txt
kafka-python
elasticsearch
fastapi
uvicorn
python-dotenv
transformers
praw
jinja2
